Avoid exporting useTableState (upcomming feature)
