+++ react/components/Table/BulkActionsSelectedRows.js
@@ -1,23 +1,12 @@
-import { Component } from 'react'
+import { memo } from 'react'
 import PropTypes from 'prop-types'
 
-class BulkActionsSelectedRows extends Component {
-  shouldComponentUpdate(nextProps) {
-    // This prevents the value from showing 0 when bar is closing
-    // It reduces the "noise" given to the user from a ux point of view
-    if (nextProps.selectedRowsLength === 0) {
-      return false
-    }
-    return true
-  }
-
-  render() {
-    return this.props.selectedRowsLength
-  }
+const BulkActionsSelectedRows = ({ selectedRowsLength }) => {
+  return selectedRowsLength
 }
 
 BulkActionsSelectedRows.propTypes = {
   selectedRowsLength: PropTypes.number.isRequired,
 }
 
-export default BulkActionsSelectedRows
+export default memo(BulkActionsSelectedRows)
