+++ react/components/Table/CheckboxContainer.js
@@ -1,43 +1,39 @@
-import React, { Component } from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'
 
 import Checkbox from '../Checkbox'
 
-class CheckboxContainer extends Component {
-  static defaultProps = {
-    partial: false,
-    disabled: false,
-  }
+const CheckboxContainer = ({ checked, partial, id, onClick, disabled }) => {
+  return (
+    <div
+      onClick={e => {
+        e.stopPropagation()
+        // prevents the onRowClick event from happening
+      }}>
+      <Checkbox
+        checked={checked}
+        partial={partial}
+        value={`${id}`}
+        id={`${id}`}
+        name={`row_${id}`}
+        onChange={() => onClick(id)}
+        disabled={disabled}
+      />
+    </div>
+  )
+}
 
-  static propTypes = {
-    checked: PropTypes.bool.isRequired,
-    partial: PropTypes.bool,
-    onClick: PropTypes.func.isRequired,
-    disabled: PropTypes.bool,
-    id: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-  }
+CheckboxContainer.defaultProps = {
+  partial: false,
+  disabled: false,
+}
 
-  render() {
-    const { checked, partial, id, onClick, disabled } = this.props
-
-    return (
-      <div
-        onClick={e => {
-          e.stopPropagation()
-          // prevents the onRowClick event from happening
-        }}>
-        <Checkbox
-          checked={checked}
-          partial={partial}
-          value={`${id}`}
-          id={`${id}`}
-          name={`row_${id}`}
-          onChange={() => onClick(id)}
-          disabled={disabled}
-        />
-      </div>
-    )
-  }
+CheckboxContainer.propTypes = {
+  checked: PropTypes.bool.isRequired,
+  partial: PropTypes.bool,
+  onClick: PropTypes.func.isRequired,
+  disabled: PropTypes.bool,
+  id: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
 }
 
 export default CheckboxContainer
