+++ react/components/Table/util/calculateTableHeight.js
