+++ react/components/Table/util/getRowHeight.js
