+++ react/components/Table/util/reducer.js
